# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "نظام المكتبة الجامعية",
        "description": "توثيق API لنظام المكتبة الجامعية",
        "contact": {
            "email": "admin@university.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/students/{id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "عرض بيانات الطالب",
                "description": "إرجاع بيانات الطالب بما في ذلك الصورة المحفوظة.",
                "operationId": "0110b7795c74a759690b8a41ad175dbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "معرف الطالب المطلوب عرضه",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم جلب بيانات الطالب بنجاح.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "يوسف طارق"
                                        },
                                        "student_id": {
                                            "type": "string",
                                            "example": "123456"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "http://127.0.0.1:8000/storage/students/image.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "تحديث صورة الطالب",
                "description": "يتم تحديث صورة الطالب عبر إرسال مسار الصورة الجديد.",
                "operationId": "3702febdfd84c8ede39e9a3f2b89a378",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "معرف الطالب المطلوب تحديث صورته",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "example": "students/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تحديث الصورة بنجاح!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Image updated successfully!"
                                        },
                                        "student": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}